+++ frb_codegen/src/library/codegen/parser/hir/flat/parser/mirror_ident.rs
@@ -1,4 +1,4 @@
-use crate::codegen::parser::mir::attribute_parser::FrbAttributes;
+use crate::codegen::parser::mir::attribute::FrbAttributes;
 use anyhow::Context;
 use itertools::Itertools;
 use proc_macro2::Ident;
